Separate out emoji widgets in chat
